rename timecard to openSpace
