+++ front/src/App/Home/Home.jsx
@@ -1,30 +1,21 @@
-import React, { useCallback } from 'react';
+import React from 'react';
 
 import { useGetAllOpenSpaces } from '#api/os-client';
 import { useUser } from '#helpers/useAuth';
-import {ChatIcon, OpenSpaceIcon, UserIcon} from '#shared/icons';
+import {ChatIcon, OpenSpaceIcon} from '#shared/icons';
 import MainHeader from '#shared/MainHeader';
 import MyGrid from '#shared/MyGrid';
 import Spinner from '#shared/Spinner';
 import { RedirectToLogin, usePushToNewOS } from '#helpers/routes';
 
 import EmptyOpenSpaces from './EmptyOpenSpaces';
-import { deleteOS } from '#api/os-client';
-import TimeCard from "#components/molecule/TimeCard.jsx";
+import OpenSpace from "#app/Home/OpenSpace.jsx";
 import markdownToTxt from "markdown-to-txt";
 
 const Home = () => {
   const pushToNewOS = usePushToNewOS();
-  const deleteOpenSpace = (osID) =>
-    deleteOS(osID).then(() => {
-      reload();
-    });
 
   const { data: openSpaces, isPending, reload: reloadOpenSpaces } = useGetAllOpenSpaces();
-  const reload = useCallback(() => {
-    reloadOpenSpaces();
-  }, [reloadOpenSpaces]);
-
   return (
     <>
       <MainHeader>
@@ -43,13 +34,8 @@ const Home = () => {
         <MyGrid columns="580px">
           {openSpaces.map((os) => (
               <>
-            {/*<OpenSpace deleteOS={() => deleteOpenSpace(os.id)} key={os.id} {...os} />*/}
-            <TimeCard
+            <OpenSpace
                 id={os.id}
-            time={{
-            start: os.startDate,
-            end: os.endDate,
-          }}
             dates={os.dates}
           title={os.name}
           description={markdownToTxt(os.description)}
